+++ app/Http/Controllers/AuthController.php
@@ -1,0 +1,82 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use Illuminate\Http\Request;
+use App\Models\User;
+use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Password;
+
+class AuthController extends Controller
+{
+    // Register
+    public function register(Request $request)
+    {
+        // Validate the request
+        $request->validate([
+            'name' => 'required|string|max:255',
+            'email' => 'required|string|email|max:255|unique:users',
+            'password' => 'required|string|min:6|confirmed',
+        ]);
+
+        // Create the user
+        $user = User::create([
+            'name' => $request->name,
+            'email' => $request->email,
+            'password' => $request->password,
+        ]);
+
+        $token = $user->createToken('auth_token')->plainTextToken;
+
+        // Return a response
+        return response()->json([
+            'message'       => 'Register berhasil',
+            'token'         => $token,
+            'user'          => $user
+        ], 201);
+    }
+
+    // Login
+    public function login(Request $request)
+    {
+        // Validate the request
+        $request->validate([
+            'name' => 'required',
+            'password' => 'required|string',
+        ]);
+
+        $user = User::where('name', $request->name)->first();
+
+        if (!$user || !Hash::check($request->password, $user->password)) {
+            return response()->json([
+                'message' => 'Login gagal, pastikan nama dan password benar'
+            ], 401);
+        }
+
+        $token = $user->createToken('auth_token')->plainTextToken;
+
+        return response()->json([
+            'message' => 'Login berhasil',
+            'token' => $token,
+            'user' => $user
+        ], 200);
+    }
+
+    // Logout
+    public function logout(Request $request)
+    {
+        $request->user()->currentAccessToken()->delete();
+
+        return response()->json([
+            'message' => 'Logout berhasil'
+        ], 200);
+    }
+
+    // User Profile
+    public function userProfile(Request $request)
+    {
+        return response()->json([
+            'user' => $request->user()
+        ], 200);
+    }
+}