+++ app/Http/Resources/Resource.php
@@ -1,0 +1,41 @@
+<?php
+
+namespace App\Http\Resources;
+
+use Illuminate\Http\Request;
+use Illuminate\Http\Resources\Json\JsonResource;
+
+class Resource extends JsonResource
+{
+    // define properties
+    public $status;
+    public $message;
+    public $resource;
+
+    /**
+     * Create a new resource instance.
+     *
+     * @param  mixed  $resource
+     * @return void
+     */
+    public function __construct( $status, $message, $resource)
+    {
+        parent::__construct($resource);
+        $this->status = $status;
+        $this->message = $message;
+    }
+
+    /**
+     * Transform the resource into an array.
+     *
+     * @return array
+     */
+    public function toArray(Request $request): array
+    {
+        return [
+            'success' => $this->status,
+            'message' => $this->message,
+            'data' => $this->resource
+        ];
+    }
+}