+++ CHANGELOG.md
@@ -18,3 +18,5 @@
 
 ### [Unreleased]
 - Create initial project
+- Add Authentication 
+- Add Schools Endpoint